splits remaining fields
