+++ test/fields/collections/Blocks/e2e.spec.ts
@@ -280,4 +280,20 @@ describe('Block fields', () => {
       })
     })
   })
+
+  describe('sortable blocks', () => {
+    test('should have disabled admin sorting', async () => {
+      await page.goto(url.create)
+      const field = page.locator('#field-disableSort > div > div > .array-actions__action-chevron')
+      expect(await field.count()).toEqual(0)
+    })
+
+    test('the drag handle should be hidden', async () => {
+      await page.goto(url.create)
+      const field = page.locator(
+        '#field-disableSort > .blocks-field__rows > div > div > .collapsible__drag',
+      )
+      expect(await field.count()).toEqual(0)
+    })
+  })
 })
