+++ test/fields/collections/Tabs/e2e.spec.ts
@@ -19,7 +19,7 @@ import { AdminUrlUtil } from '../../../helpers/adminUrlUtil.js'
 import { initPayloadE2ENoConfig } from '../../../helpers/initPayloadE2ENoConfig.js'
 import { reInitializeDB } from '../../../helpers/reInitializeDB.js'
 import { RESTClient } from '../../../helpers/rest.js'
-import { TEST_TIMEOUT_LONG } from '../../../playwright.config.js'
+import { POLL_TOPASS_TIMEOUT, TEST_TIMEOUT_LONG } from '../../../playwright.config.js'
 import { tabsFieldsSlug } from '../../slugs.js'
 
 const filename = fileURLToPath(import.meta.url)
@@ -131,5 +131,27 @@ describe('Tabs', () => {
     await expect(page.locator('#field-tab__array__0__text')).toHaveValue(
       "Hello, I'm the first row, in a named tab",
     )
+  })
+
+  test('should save preferences for tab order', async () => {
+    await page.goto(url.list)
+
+    const firstItem = page.locator('.cell-id a').nth(0)
+    const href = await firstItem.getAttribute('href')
+    await firstItem.click()
+
+    const regex = new RegExp(href.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'))
+
+    await page.waitForURL(regex)
+
+    await page.locator('.tabs-field__tabs button:nth-child(2)').nth(0).click()
+
+    await page.reload()
+
+    const tab2 = page.locator('.tabs-field__tabs button:nth-child(2)').nth(0)
+
+    await expect(async () => await expect(tab2).toHaveClass(/--active/)).toPass({
+      timeout: POLL_TOPASS_TIMEOUT,
+    })
   })
 })
